working on tests and fixing the domain structure
