+++ pyspectrum/__init__.py
@@ -1,6 +1,4 @@
 from .core import Spectrum, Peak, Domain
-#from .identification import FindPeaksDomain, FindPeaksCenters, Convolution, gaussian_2_dev
-#from .fitting import GaussianWithBGFitting
-#from .calibration import Calibration, germanium_fwhm, standard_fwhm_generator
-#from .peak import Peak
-#from .io import TimeChannelParser
+from .io import TimeChannelParser
+from .calibration import ResolutionCalibration, AxisCalibration
+
