+++ pyspectrum/background/__init__.py
@@ -1,1 +1,3 @@
 from .snip import SNIPBackground
+from .polynomial import IterativePolyFit, IterativePolyFitWithMinimum
+from .minimum_in_envolope import MinimaEnvelopeBackground
