+++ pyspectrum/domain_analysis/background.py
@@ -1,7 +1,7 @@
 import numpy as np
 from scipy.special import erf
 from pyspectrum.core.domain import Domain
-from .morphology import find_domain_peaks, domain_bases
+from pyspectrum.domain_analysis.characterization.morphology import find_domain_peaks, domain_bases
 
 def domain_erf_background(domain: Domain,
                           smooth=True,
