+++ pyspectrum/domain_analysis/characterization/moment.py
