+++ pyspectrum/domain_analysis/fitter/__init__.py
@@ -1,3 +1,3 @@
-""" A module for fitting objects. \n The fitting objects inherit from PeakFit"""
+""" A module for fitter objects. \n The fitter objects inherit from PeakFit"""
 from .abstract_fitting_class import PeakFit
 from .std_gaussian_fitting import GaussianWithBGFitting
