+++ pyspectrum/domain_analysis/fitter/skew_gauss.py
