+++ pyspectrum/identification/__init__.py
@@ -1,2 +1,3 @@
 from .convolution import Convolution
+from .snr import SNRFinder
 from pyspectrum.identification.kernels import gaussian_2_dev
