--- pyspectrum/identification/cwt.py
@@ -1,22 +1,0 @@
-"""
-===========================================================
-⚠️  DEVELOPMENT / EXPERIMENTAL MODULE
-===========================================================
-
-This module is under active development.
-
-- The API is NOT stable
-- Do NOT rely on this for production analysis
-
-Intended purpose:
------------------
-Method to find domains with peaks
-
-Planned changes:
-----------------
-- Writing the function
-
-Author: Achiya
-Status: DEVELOPMENT
-===========================================================
-"""
