+++ requirements.txt
@@ -1,8 +1,6 @@
-setuptools~=78.1.1
-xarray~=2023.10.1
-numpy~=1.26.0
-uncertainties~=3.1.7
-pandas~=2.1.1
-lmfit~=1.2.2
-scipy~=1.11.3
-matplotlib~=3.7.2
+xarray>=2022.6
+numpy>=1.22.0, <3.0
+uncertainties>=3.1
+pandas>=2.0.0, <4.0
+scipy>=1.8, <2.0
+setuptools>=70.0.0
