+++ setup.py
@@ -1,10 +1,10 @@
-from setuptools import setup
+from setuptools import setup, find_packages
 
 setup(
     name='PySpectrum',
     version='0.2',
-    packages=['pyspectrum', 'pyspectrum.peak', 'pyspectrum.spectrum', 'pyspectrum.calibration',
-              'pyspectrum.fitting', 'pyspectrum.io', 'pyspectrum.identification'],
+    python_requires='>=3.9',
+    packages=find_packages(),
     url='https://github.com/achiyaAmrusi/pySpectrum',
     license='MIT license',
     author='Achiya Yosef Amrusi',
